Implement login and more
Main work here is a LoginScreen which works pretty much the same as
login did in the Flutterv sample. It uses localStorage to remember the
user email and client ID; allowing login/logout while offline.

Restructured the code slightly to have the Replicache object be part of
the state too.