Use postinstall to download binaries
The downloaded files are put in `bin/`. The real files are downloaded
when someone does `npm install`. To make npm happy we put dummy files in
`bin/`. Also, to ensure that we do not accidentally publish the binaries
as part of the npm package I added a little script that looks at the
size of those files.

Also set up `package.json` `bin` fields to allow invoking the files
using `npx` instead of the full path into
`node_modules/replicache/bin/`.