Implement basic transaction support in dispatch (#71)
Creates a new spawned future for each database connection, and causes dispatch_loop()
to forward messages for that database to that future. There, the db::Store and a map of transactions
are owned on the stack, and a FuturesUnordered is used to enqueue a future per incoming RPC to
run in parallel while depending on the stack for ownership of the transactions.

With this, multiple reads can run in parallel on a transaction, and multiple write transactions will queue
until they can receive a write lock from the underlying Store.