Switch to Web Test runner
https://modern-web.dev/docs/test-runner/overview/

Now we are running the tests in real browsers so we can use real modules
and real DOM.

This also removes the concept of running the tests with the test server.
The actual test server parts will be remove in a follow up.

Update GH Action to support Playwright. We run the test in Chromium and
in Firefox. Webkit seems to have some TLS issues when running as a GH
action :'(

Fixes #75