Add performance CI
https://medium.com/@thomaspoignant/ci-build-performance-testing-with-github-action-e6b227097c83
https://github.com/marketplace/actions/continuous-benchmark

We have a custom runner that uses Playwright. This used `sample/perf` as a
base for the new perfomance test. It outputs format that looks similar
enough to BenchmarkJS to allow the Github Action to extract the data.

The data is `git push`ed to the `dev-data` branch. We display the data
using https://raw.githack.com/

Fixes https://github.com/rocicorp/repc/issues/178