Service Worker
This moves the wasm part of Replicache to a service worker.

This depends on https://github.com/rocicorp/repc/pull/275

Right now the marshal between main and worker is naive and uses a 1 to 1
mapping (more or less) to the wasm `dispatch` function.

There is a lot of follow up to deal with after this:
- Use MessageChannel for speed.
- Make scan progressive
- Broadcasting changes to clients
- Versioning
- Mutliple tabs closing the same database
- ...

Towards #143
Towards #226