Call subscribe onData when query retval changes
For `subscribe`, we now only call `onData` if the return value of the
`body` function changes compared to last time.

This requires that we store the last return value to be able to compare
with that.

We consider the values equal if they behave like `JSON.stringify(a) ===
JSON.stringify(b)` except for the ordering of the property keys.

*BREAKING CHANGE*

This is a breaking API change because subscribe now has a slightly more
restricted type which can lead to new TypeScript type errors.

Also, we changed the type signature from `ToJSON` to `JSONValue` in a
few places.

https://github.com/rocicorp/replicache-sdk-js/issues/63