ScanIterator: Get rid of stream controller
At this point and for the foreseeable future, the actual iteration is
synchronous so enqueueing using a ReadableStreamDefaultController is
overkill (it will just put all the entries in an internal list).

We might want to allow some other way of iterating over the scan result
that allows early exits but that requires a change in the API so we can
deal with that later.

Follow up to #235