Fix: compare to current main commit when computing changes from pull.
When computing changed keys for a pull, we were comparing the last snapshot on main to the current state of the sync branch.

This is not correct, because main will typically have subsequent pending changes which have not been confirmed by server yet. We need to consider these changes too when computing the diff.

Fixes #466