Respond to all code review feedback.
Also remove explicit `window.` in many places, and add a `withTimeout()`
utility to wrap a promise with a timeout. I used this for both new tab
initialization and calls to the new tab as working on tests is abysmal
if one makes a small error with the default 30s timeout, and this seemed
to help.a lot.