Add utility to open a window and run code in it for tests. (#492)
Use this to start a test for navigator.locks, verifying that it works on
all browsers we care about.

When test-util.ts newTab() is called, a new browser window is opened,
with all console logs from it funnelled back to the main test process. A
module (usually that of the current test) may be loaded and all exported
symbols made available, allowing type-checked Typescript code to be
executed easily. The run() method on the resulting tab may then be
called to execute snippets of code, allowing for testing of interaction
between two windows at the same origin.

Example:

```
import {newTab} from './test-util';

test('storage', async () => {
  const tab = await newPage();
  await tab.run('window.localStorage.setItem("test", "modified")');
  expect(window.localStorage.getItem('test')).to.equal('modified');
}
```