Vary keys written to localStorage to avoid races.
I can't explain exactly why this made a difference, but perhaps some
browsers elide some storage events if multiple writes happen to the same
key close together in time? I'm not sure, but this seems much more
reliable when running all tests on my computer.

I also noted that the dynamic import calls takes hundreds of
milliseconds in some cases, particularly on Firefox, so this allows for
a longer timeout for cases when this is even slower when running lots of
tests on all three browsers at once.