Feat: Use ReadonlyJSONValue in read transactions
BREAKING CHANGE: If you are using TypeScript you might need to update
your type annotations to use ReadonlyJSONValue instead of JSONValue in a
few places.

ReadTransaction return the underlying value as ReadonlyJSONValue. This
is because you should not mutate these since they may be cached
internally.

WriteTransaction returns a clone of the value and is safe to mutate.

WriteTransaction `put` clones the value to ensure that internal valie is
not mutated.

Fixes #480