refactor: Simplify Prolly Map (#518)
There was actually a bug in binary search as well. Doh!

There is a slight change in semantics now. The changed keys now includes
no-op puts. In other words we do not check for equality with the old
value. This leads to a possibly larger changed keys but which might lead
to the subscription body being run more often than needed.

Fixes #481