chore: Add perf tests for hashing
This tests a few things but most importantly, it tests the performance
of native hashing vs wasm.

Another small win that is shown in this test is to not use an Uint8Array
with utf8 but instead use a Uint16 with utf16. This is because JS
strings are already in UTF16 so there is no need to convert them.