B+ Tree (#610)
This adds a B+Tree that we use instead of the prolly tree. The old prolly tree was not actually using a prolly tree but a flat array of entries. That meant that we read and wrote all the data on every transaction which of course is not very fast.

With a B+Tree we only load the parts of the Map that you are interested in.

The current target chunk size is 8kb (min is 8kb and max is 16kb)

Fixes #596