feat!: Expose clientID on ReadTransaction
Exposes clientID on ReadTransaction (and subclasses, e.g. WriteTransaction, IndexTransaction, etc).

This change was previously reverted due to it causing perf test regressions.  The
underlying cause of those regressions was addressed in commit cc2bbeb.

BREAKING CHANGE: The Replicache class no longer duck-types as a ReadTransaction (prior to
52bab5d Replicache explicitly implemented ReadTransaction).  Any place where Replicache
is assigned or passed as a ReadTransactio will need to be udpated, most likely to use
Replicache.query.

Closes #541