refactor: Use opaque Hash type
This introduces an opeaque hash type. Now it is an error to pass a
string into a function that requires a Hash. It is also an error to pass
a Hash into a function that requires a string.

To go from a string to Hash, use `parse`.

To go from a Hash to a string do `String(hash)` (or `hash.toString()`,
`hash + ''` etc).