refactor: Change commit refs computation (#675)
Main motivation is to be able to compute the refs directly from the
chunk data but this simplifies the code a bit. We used to have an object
telling us if the Hash in the commit was Weak or Strong. It was only
used internally when computing the actual refs. Now we keep this logic
in one place and we do need to keep this extra type around.