refactor: Get rid of meta key
We currently store the data in `c/<hash>/d` and the refs this data
references in `c/<hash>/m` (m for meta). The idea was that we would be
able to read the refs without loading the data, but in practice this
does not matter much, especially not in the MemStore. The only case
where we could really skip reading the data key was when we do GC.

With this patch we do not store the refs in a seperate key. Instead, we
store the type of the chunk in the data key `[type: ChunkType, data:
JSONValue]`. Given the `ChunkType` we can then get the refs from the
data directly.

This does mean that GC will read the B+Tree leaf node data. For MemStore
this should not be a problem, but if it is we can avoid it by checking
the level of the B+Tree node and not read leaf node during GC.