refactor: Add PersistGatherVisitor
Add a DB/Dag Visitor -- This walks the entire dag using a semantic
visitore, which knows what each chunk represents.

Then implement the PersistGatherVisitor as a visitor of the Dag Visitor
which stops the traversal when it finds a non temp hash. It collects all
the chunks it sees and exposes them as a property.

Towards #671