refactor: Optimize scan
There are two kinds of optimizations in here:

1. Get rid of intermediate for await loops.
2. Get rid of yield*

Both of these adds extra Promise and IteratorResult objects.

By passing the convertEntry function all the way down into the BTree
iterator we do not need the intermediate for await loops.

In a few places we can return the async iterable iterator instead of
yield* it. This only works if the function/method is not `async` of
`async *`.

Towards #711