feat: A transformer that computes the chunk hash
This does not need a dag.Read or dag.Write. It only operates on the
gathered chunks in the map from the previous step.

The input is a `Map<TempHash, Chunk<TempHash>>` and the output is the
same logical map but the hashes have been computed based on the chunk
data; `Map<PerHash, Chunk<PerHash>>`

Towards Implement Dueling Dags #671