feat!: Include versions in IDB name (#741)
For Simplified Dueling Dags we need to ensure that different tabs
running different versions of Replicache do not interact with IDB data
it does not know how to read/write.

To achieve this the name if the IDB database now contains the
`REPLICACHE_FORMAT_VERSION` (which is currently at `3`).

The IDB name also contains the `schemaVersion` som if the schema changes
a fresh IDB database is used. The motivation is the same. Multiple tabs
with different schemaVersions should not interact with the same IDB
database.

BREAKING CHANGE

Towards #671