refactor: Make commit statics module functions (#749)
Static methods are generally an anti-pattern in JS. They are sometimes
nice from an API perspective, but tree shaking generally has problems
with them.

The only real valid use case I can think of is when you need to inherit
static methods. In other words when your statics references `this`.