feat: Add persist to Replicache
See Simplified Dueling Dags design doc

Replicache now creates two DAG stores backed by two different KV stores.
These are referred to as `memdag` and `perdag`.

Replicache operates on the memdag and once in a while it does a persist
which syncs data from the memdag to the perdag.

Towards #671