fix: Improve isolation of tests' and benchmarks' indexeddb state to reduce flakiness (#822)
In tests add a uuid to indexeddb database names (Replicache name and IDBDatbasesStore DB) 
to isolate tests' indexed db state.
 
Also fixes a bug in kv.IDBStore which was blocking IndexedDB opened by these stores from being 
deleted.   In order to not block deletion of the db, the connection needs to be closed on 
`onversionchange`.  Previously the code was only setting up 
`db.onversionchange = () => db.close()` in `onupgradeneeded` which only fires if the db didnt 
already exist.  Code is updated to always setup `db.onversionchange = () => db.close()`. 

While this fix allows the IndexedDB databases to be reliably deleted, it did not prevent races 
around deletion.  Before isolating with uuid, I was observing that after one test's teardown 
await the deletion of a database with name X, if the next test opened the database with name 
X, the test would _sometimes_ get an error that its connection to X was closed, suggestion some 
race where the deletion is not truly complete when the success callback for the deletion is invoked.

Fixes #819