fix: Fix flakiness of replicache-persist.test.ts on Webkit (#828)
**Problem**
replicache-persist.test is flacky on webkit because the persist process does not always complete before we 
create a new replicache and try to read the persisted data.  This happens more on webkit because it uses a 
timeout rather than request idle callback to start persist.

**Solution**
Wait for persist to complete (detected by polling the ClientMap) before creating a new Replicache and verifying
it bootstraps from the persisted data.