fix: Catch errors in background interval processes and log appropriatly. (#846)
Problem
=======
We have a report of the following exception being thrown from `heartbeat.ts` after a Replicache instance is closed and a new one created as part of a development setup using Replicache, Reach hooks, and Next with HMR.

```
DOMException: Failed to execute 'transaction' on 'IDBDatabase': The database connection is closing.
```

We do stop the heartbeat interval when a Replicache instance is closed, however there is a race that can lead to the above exception: if the Replicache instance is closed while the hearbeat update is running.  

This is a fairly narrow race, so I'm still uncertain if this is what the issue reporter is hitting. 

Solution
=======
Catch errors in interval based background processes and log them to 'debug' if the error occurred after the Replicache instance was closed (as this is an expected error), and to 'error' otherwise.  Applied this to the "heartbeat" and "ClientGC" processes.  The "mutation recovery" process already does this.

Also added so additional debug logging to aid in further debugging if this does not fix the issue for the reporter.


Fixes #843