feat: Check if client exists in persist
We now check if the client ID exists in the client map when we do a
`persist`. If it doesn't we throw a `MissingClientError`.

For testing purpuse we can skip this check.

The intended use is to handle clients that are missing and raise an
"event" on the Replicache instance when this happens.

Towards #784