feat!: Switch to use a ScanReader and expose it to the API.
This changes scan to use a ScanReader instead of an async iterator.

It also exposes a function that returns a ScanResult from a ScanReader
and ScanOptions.

```ts
declare const options: ScanOptions;
declare const reader: ScanReader;

const scanResult = makeScanResult(reader, options);
```

If you are trying to implement Replicache's scan API you now only need
to write a function that creates a ScanReader.
makeScanResult will take care of seeking to the correct
position and reading the data.

When working with index scans you need to use `ScanReader<IndexKey>`

Closes #607