feat!: Expose a way to reuse ScanResult
We now expose a method called `makeScanResult`. It takes a `ScanOptions`
and a function that returns an async iterator.

```ts
makeScanResult({prefix: 'b'}, async function* (fromKey) {
  // yield ['a', 1];
  yield ['b', 2];
});
```

or when using an index:

```ts
makeScanResult(
  {prefix: 'b', indexName: 'i'},
  async function* (indexName, fromSecondaryKey, fromPrimaryKey) {
    // yield [['as', 'ap', 1];
    yield [['bs', 'bp', 2];
});
```

To make this work we moved the limit and exclusive handling to the top
level iterator loop.

We now compute the fromKey and pass that into the iterator.

Fixes #607