refactor: Convert the key for btree index iterator early (#932)
Instead of converting the key from third party BTree iterators to our
encoded string, convert our encoded string to to an entry. Then let the
main scan loop work with IndexKey as needed.

The benefit is that for external iterators we do not have to go from
IndexKey to string and back to IndexKey.